Windows fixes part 1

+++ decoder_modules/ch_extravhf_decoder/src/dsp/dsd_demod.cpp
@@ -327,7 +327,7 @@ namespace dsp {
         int DSD::getSymbol (int have_sync) {
             int i, sum, symbol, count;
             short sample;
-            ssize_t result;
+            int result;
 
             sum = 0;
             count = 0;
